several more styling changes and using api/users
