+++ package.json
@@ -8,7 +8,6 @@
     "@testing-library/user-event": "^12.6.2",
     "axios": "^0.21.1",
     "dotenv": "^8.2.0",
-    "firebase": "^8.2.5",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
