+++ src/components/HomeView/SearchBody/SearchBody.css
