+++ src/components/SharedComponents/Card.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Card = () => (
+  <div />
+);
+
+export default Card;
