+++ src/index.jsx
@@ -2,7 +2,7 @@ import 'dotenv/config';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './views/App/App';
+import App from './components/App/App';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
