--- src/views/HomeView/MainFooter/MainFooter.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import './MainFooter.css';
-
-const MainFooter = () => (
-  <div className="footer-container">
-    <div className="footer-content">
-      <p>Created by F Rodriguez</p>
-    </div>
-  </div>
-);
-
-export default MainFooter;
