--- src/views/SharedComponents/components/Button.jsx
@@ -1,16 +1,0 @@
-import React, { Children } from 'react';
-import './Button';
-
-const Button = ({buttonContent, clickHandler, children}) => {
-    return(
-        <div className="button-container">
-            <button
-                className="custom-button"
-                onClick={clickHandler}>
-                {children}
-            </button>
-        </div>  
-    );
-};
-
-export default Button;
