--- src/views/SharedComponents/components/Card.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Card = () => {
-    return (
-        <div></div>
-    );
-};
-
-export default Card;
