--- src/views/SharedComponents/components/TextInput.css
@@ -1,7 +1,0 @@
-.custom-text-input-container {
-    background: blue;
-}
-
-.custom-text-input {
-    
-}
